requirements.txt
# PYLON - SKT Network센터 Energy Operations Platform
# Python 3.9+ 권장

# Core framework
streamlit>=1.30.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # Parquet support

# Visualization
plotly>=5.18.0

# Utilities
python-dateutil>=2.8.2
PyYAML>=6.0  # Configuration files

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

